Implement remove book button for each book
I saw two options here, one could have been
to simply delete the element in the array,
leaving undefined which would have kept the index
correctly. I went for option two and choose
to simply reprint to the dom every time a
book is removed, thus rewriting the data-library-index
to match the new collection. Maybe a combination
of both could be interesting in a real project
like using undefined until a certain point and
rewrite one in a while. This would make it efficient
but still clean.